requirements.txt
streamlit
opencv-python-headless
pillow
pytesseract
numpy
